EN-62: enlarge card info + refactor
