+++ src/employees.js
@@ -4,53 +4,46 @@ import {
   DateField,
   List,
   TextField,
-  Show,
-  TabbedShowLayout,
-  SimpleShowLayout,
-  FunctionField,
   RecordContextProvider,
-  ArrayField,
   ShowButton,
   EditButton,
-  Tab,
   TopToolbar,
   CreateButton,
   ExportButton,
   useListContext,
-  ReferenceArrayField,
-  SingleFieldList,
-  ChipField,
-  ReferenceManyField,
-  ReferenceField,
-  NumberField,
   SearchInput,
-  useGetRecordId,
 } from "react-admin";
 import {
   Card,
   CardContent,
-  CardMedia,
   CardActions,
   Typography,
+  CardActionArea,
+  
 } from "@mui/material";
-import { Avatar, Box, Divider, Grid } from "@mui/material";
+import { 
+  red,
+  blueGrey,
+  pink,
+  deepPurple, 
+  indigo, 
+  blue,
+  cyan,
+  teal, 
+  green,
+  orange
+} from '@mui/material/colors';
+import { Avatar, Box, Grid } from "@mui/material";
 import CreateForm from "./components/forms/CreateForm";
 import EditForm from "./components/forms/EditForm";
-import RedirectButton from "./components/RedirectButton";
-import { HasPermissions, ListActions } from "./components/layout/CustomActions";
+import { HasPermissions } from "./components/layout/CustomActions";
 import RowRadioButtonGroup from "./components/buttons/RowRadioButtonGroup";
 import { useState } from "react";
+import { Link } from "react-router-dom";
 
-const COLOR_green = "#efe";
-const COLOR_white = "#white";
-
-const activeContractRowStyle = (record) => ({
-  backgroundColor: record.active === true ? COLOR_green : COLOR_white,
-});
-
-const DisplayRecordCurrentId = () => {
-  return useGetRecordId();
-};
+const COLOR_BG = [red[500], blueGrey[500], pink[500], deepPurple[500], 
+                  indigo[500], blue[500], cyan[500], teal[500], green[500], 
+                  orange[500]]
 
 const formData = [
   {
@@ -120,13 +113,13 @@ const formData = [
 const employeeFilters = [<SearchInput source="q" alwaysOn />];
 
 export const EmployeeList = () => {
-  const [radioValue, setRadioValue] = useState("card");
+  const [viewOptionValue, setRadioValue] = useState("card");
 
   const handleChange = (event) => {
     setRadioValue(event.target.value);
   };
 
-  const radioButtonOptions = [
+  const viewOptions = [
     {
       id: 1,
       label: "Card",
@@ -156,9 +149,9 @@ export const EmployeeList = () => {
           <Box>
             <RowRadioButtonGroup
               title={"View mode"}
-              value={radioValue}
+              value={viewOptionValue}
               handleChange={handleChange}
-              options={radioButtonOptions}
+              options={viewOptions}
             />
           </Box>
           <Box>
@@ -166,7 +159,7 @@ export const EmployeeList = () => {
             <ExportButton />
           </Box>
         </TopToolbar>
-        <EmployeeInformation renderAs={radioValue} />
+        <EmployeeInformation renderAs={viewOptionValue} />
       </>
     </List>
   );
@@ -177,50 +170,67 @@ const EmployeeInformation = ({ renderAs = "list" }) => {
   if (isLoading) {
     return null;
   }
+
   if (renderAs === "card") {
     return (
-      <Grid container spacing={1} sx={{ marginTop: "1em" }}>
-        {data.map((record, index) => (
-          <RecordContextProvider key={index} value={record}>
-            <Grid xs={2} item>
-              <Card sx={{ maxWidth: 300 }}>
-                {/*Profile image hardcoded until photo upload feature is in palce*/}
-                <CardMedia
-                  component="img"
-                  height="200"
-                  image="https://entropay-assets.s3.us-east-1.amazonaws.com/default-profile.png"
-                />
-
+      <Grid container spacing={2} sx={{ marginTop: "1em" }}>
+      {data.map((record, index) => (
+        <RecordContextProvider key={index} value={record}>
+          <Grid xs={2} item>
+            <Card>
+              <CardActionArea component={Link} to={`${record.id}` + "/show"} style={{ textDecoration: 'none' }}>
                 <CardContent sx={{ padding: 1 }}>
-                  <Typography variant="h5" component="h5" align="center">
-                    {`${record.firstName} ${record.lastName}`}
-                  </Typography>
-                  <Typography align="center">{record.personalEmail}</Typography>
-                  <CardActions>
-                    <ShowButton />
-                    {HasPermissions("employees", "update") && <EditButton />}
-                  </CardActions>
+                  <Box sx={{ display: 'flex', justifyContent: 'center' }}>
+                    <Avatar
+                      alt="Employee"
+                      sx={{ 
+                        width: 100, 
+                        height: 100, 
+                        bgcolor: COLOR_BG[`${record.internalId}`.charAt(`${record.internalId}`.length - 1)], 
+                        fontSize: 50, 
+                        margin: 2
+                      }}
+                    >{`${record.firstName}`.charAt(0)}{`${record.lastName}`.charAt(0)}
+                    </Avatar>
+                  </Box>
+                  <Box sx={{ minHeight: 155}}>
+                    <Typography noWrap variant="h5" component="h5" align="center">
+                      {record.firstName} {record.lastName}
+                    </Typography>
+                    <Typography noWrap align="center">{record.personalEmail}</Typography>
+                    <Typography noWrap align="center"><DateField source="startDate"/></Typography>
+                    <Typography noWrap align="center">{record.state} / {record.country}</Typography>
+                    <Typography noWrap align="center">{record.client} / {record.project}</Typography>
+                    <Typography noWrap align="center">{record.role}</Typography>
+                  </Box>
                 </CardContent>
-              </Card>
-            </Grid>
-          </RecordContextProvider>
-        ))}
-      </Grid>
+              </CardActionArea>
+              <CardActions>
+                <ShowButton />
+                {HasPermissions("employees", "update") && <EditButton />}
+              </CardActions>
+            </Card>
+          </Grid>
+        </RecordContextProvider>
+      ))}
+    </Grid>
+    );
+  } else {
+    return (
+      <Datagrid rowClick="edit">
+        <TextField source="internalId" />
+        <TextField source="firstName" />
+        <TextField source="lastName" />
+        <TextField source="personalEmail" />
+        <TextField source="startDate" />
+        <TextField source="city" />
+        <TextField source="country" />
+        <TextField source="client" />
+        <TextField source="project" />
+        <TextField source="role" />
+        </Datagrid>
     );
   }
-  return (
-    <Datagrid rowClick="edit">
-      <TextField source="internalId" />
-      <TextField source="firstName" />
-      <TextField source="lastName" />
-      <TextField source="personalEmail" />
-      <TextField source="mobileNumber" />
-      <TextField source="address" />
-      <TextField source="city" />
-      <TextField source="country" />
-      <TextField source="personalNumber" />
-    </Datagrid>
-  );
 };
 
 export const EmployeeEdit = () => (
@@ -229,196 +239,4 @@ export const EmployeeEdit = () => (
 
 export const EmployeeCreate = () => (
   <CreateForm formData={formData} title="Employees" resource="employees" />
-);
-
-export const EmployeeProfile = () => (
-  <Show
-    title="Show employee"
-    actions={<ListActions entity={"employees"} />}
-    emptyWhileLoading
-  >
-    <Grid
-      container
-      direction="row"
-      justifyContent="flex-start"
-      alignItems="center"
-    >
-      <Grid item>
-        <Box m={2}>
-          {/*Profile image hardcoded until photo upload feature is in palce*/}
-          <Avatar
-            alt="Employee"
-            src="https://entropay-assets.s3.amazonaws.com/default-profile.png"
-            sx={{ width: 100, height: 100 }}
-          />
-        </Box>
-      </Grid>
-      <Grid item>
-        <SimpleShowLayout divider={<Divider flexItem />}>
-          <FunctionField
-            label=""
-            render={(record) => `${record.firstName} ${record.lastName}`}
-          />
-          <TextField label="" source="personalEmail" />
-        </SimpleShowLayout>
-      </Grid>
-      {/*
-      <Grid item>
-        <SimpleShowLayout divider={<Divider flexItem />}>
-          <TextField label="Current assigment" source="" />
-          <DateField label="Hired Date" source="" />
-        </SimpleShowLayout>
-      </Grid>
-      Hidden empty fields until developed
-      */}
-    </Grid>
-    <TabbedShowLayout>
-      <Tab label="Personal and financial information">
-        <Grid
-          container
-          direction="row"
-          justifyContent="flex-start"
-          alignItems="flex-start"
-        >
-          <SimpleShowLayout divider={<Divider flexItem />}>
-            <TextField source="internalId" label="Internal ID" />
-            <TextField source="taxId" />
-            <TextField source="phoneNumber" />
-            <TextField source="mobileNumber" />
-            <TextField source="address" />
-            <TextField source="city" />
-            <TextField source="emergencyContactFullName" />
-            <TextField source="emergencyContactPhone" />
-            <TextField source="notes" />
-          </SimpleShowLayout>
-          <SimpleShowLayout divider={<Divider flexItem />}>
-            <TextField source="state" />
-            <TextField source="zip" />
-            <TextField source="country" />
-            <DateField source="birthDate" />
-            <TextField source="personalNumber" />
-            <TextField source="healthInsurance" />
-            <ReferenceArrayField
-              label="Profile"
-              reference="roles"
-              source="profile"
-            >
-              <SingleFieldList>
-                <ChipField source="name" />
-              </SingleFieldList>
-            </ReferenceArrayField>
-            <ReferenceArrayField
-              label="Technologies"
-              reference="technologies"
-              source="technologies"
-            >
-              <SingleFieldList>
-                <ChipField source="name" />
-              </SingleFieldList>
-            </ReferenceArrayField>
-          </SimpleShowLayout>
-        </Grid>
-        <ArrayField source="paymentInformation">
-          <Datagrid>
-            <TextField source="platform" />
-            <TextField source="country" />
-            <TextField source="cbu" label="Alias/CBU" />
-          </Datagrid>
-        </ArrayField>
-      </Tab>
-      <Tab label="Contracts">
-        <ReferenceManyField
-          label=""
-          reference="contracts"
-          target="employeeId"
-          sort={{ field: "startDate", order: "DESC" }}
-        >
-          {HasPermissions("contracts", "create") && (
-            <RedirectButton
-              form="create"
-              resource="contracts"
-              text="+ CREATE"
-              recordId={DisplayRecordCurrentId()}
-              source="employeeProfile"
-            />
-          )}
-          <Datagrid rowStyle={activeContractRowStyle}>
-            <ReferenceField
-              source="contractType"
-              reference="contracts/contract-types"
-            >
-              <ChipField source="value" />
-            </ReferenceField>
-            <FunctionField
-              label="Status"
-              sortBy="active"
-              sortByOrder="ASC"
-              render={(record) =>
-                record.active === true ? "Active" : "Inactive"
-              }
-            />
-            ;
-            <ReferenceField source="companyId" reference="companies">
-              <TextField source="name" />
-            </ReferenceField>
-            <DateField source="startDate" />
-            <DateField source="endDate" />
-            <ReferenceField source="roleId" reference="roles">
-              <ChipField source="name" />
-            </ReferenceField>
-            <ReferenceField source="seniorityId" reference="seniorities">
-              <ChipField source="name" />
-            </ReferenceField>
-            <NumberField source="hoursPerMonth" />
-            <NumberField source="vacations" />
-            <TextField source="benefits" />
-            <TextField source="notes" />
-            {HasPermissions("contracts", "update") && <EditButton />}
-          </Datagrid>
-        </ReferenceManyField>
-      </Tab>
-      <Tab label="Assigments">
-        <ReferenceManyField
-          label="Assignments"
-          reference="assignments"
-          target="employeeId"
-          sort={{ field: "startDate", order: "DESC" }}
-        >
-          {HasPermissions("assignments", "create") && (
-            <RedirectButton
-              form="create"
-              resource="assignments"
-              text="+ CREATE"
-              recordId={DisplayRecordCurrentId()}
-              source="employeeProfile"
-            />
-          )}
-          <Datagrid>
-            <ReferenceField source="projectId" reference="projects">
-              <TextField source="name" />
-            </ReferenceField>
-            <DateField source="startDate" />
-            <DateField source="endDate" />
-            <ReferenceField source="roleId" reference="roles">
-              <ChipField source="name" />
-            </ReferenceField>
-            <NumberField source="hoursPerMonth" />
-            <TextField source="billableRate" />
-            <ReferenceField source="currency" reference="contracts/currencies">
-              <TextField source="name" />
-            </ReferenceField>
-            <TextField source="labourHours" />
-            <ReferenceField source="seniorityId" reference="seniorities">
-              <ChipField source="name" />
-            </ReferenceField>
-            {HasPermissions("assignments", "update") && <EditButton />}
-          </Datagrid>
-        </ReferenceManyField>
-      </Tab>
-      {/*<Tab label="Vacations and Licencies"></Tab>
-      <Tab label="Documents"></Tab>
-      Hidden empty tabs until developed
-      */}
-    </TabbedShowLayout>
-  </Show>
 );
