+++ src/resources.js
@@ -9,8 +9,10 @@ import {
     EmployeeCreate,
     EmployeeEdit,
     EmployeeList,
-    EmployeeProfile,
 } from "./employees";
+import {
+    EmployeeProfile,
+} from "./employeeProfiles";
 import { CompanyList, CompanyEdit, CompanyCreate } from "./company";
 import { ContractList, ContractEdit, ContractCreate } from "./contracts";
 import { SeniorityList, SeniorityEdit, SeniorityCreate } from "./seniorities";
